add logo to app bar
