+++ package.json
@@ -28,7 +28,8 @@
     "mongoose": "^6.8.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.6.1"
+    "react-router-dom": "^6.6.1",
+    "typescript-plugin-css-modules": "^4.1.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.15",
