+++ tsconfig.json
@@ -100,5 +100,5 @@
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true /* Skip type checking all .d.ts files. */
   },
-  "include": ["src/", "client/globals.d.ts"]
+  "include": ["src/*", "src/globals.d.ts"]
 }
